Fixes princialmente relacionados com as Doações

+++ frontend/app/components/entities/DeleteModal.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 import { useApiStore } from "@/stores/api"
+import {useAuthStore} from "~/stores/auth";
 
 const api = useApiStore()
 const toast = useToast()
@@ -17,7 +18,18 @@ const openModel = computed({
 })
 
 const onSubmit = async () => {
+  if (!useAuthStore().hasPermission('ENTITIES_DELETE')) return;
+
   if (!props.id) return
+
+  if (!await checkServerAccess()) {
+    useToast().add({
+      title: 'Sem ligação à internet!',
+      description: 'Não é possível guardar alterações sem estar ligado à internet. Tente novamente mais tarde',
+      color: 'error'
+    });
+    return;
+  }
 
   try {
     await api.deleteEntity(props.id)
