+++ frontend/app/components/users/DeleteModal.vue
@@ -1,5 +1,7 @@
 <script setup lang="ts">
 
+import {useAuthStore} from "~/stores/auth";
+
 const api = useApiStore()
 const toast = useToast()
 const props = defineProps<{
@@ -16,7 +18,18 @@ const openModel = computed({
 })
 
 const onSubmit = async () => {
+  if (!useAuthStore().hasPermission('USERS_DELETE')) return;
+
   if (!props.id) return
+
+  if (!await checkServerAccess()) {
+    useToast().add({
+      title: 'Sem ligação à internet!',
+      description: 'Não é possível guardar alterações sem estar ligado à internet. Tente novamente mais tarde',
+      color: 'error'
+    });
+    return;
+  }
 
   try {
     await api.deleteUser(props.id)
